requirements.txt
librosa
numpy
scikit_learn
tensorflow
pyYAML
pytest
